% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Barley}
\alias{Barley}
\title{Barley}
\format{
Data for a total of 32 samples of yield of barley
\describe{
  \item{Yield}{yield of barley pr. 1000 square meter}
  \item{Variety}{two varieties of barley (‘C1’ and ‘C2’)}
  \item{Soil}{two soil types (‘clay’ and ‘sand’)}
  \item{Fertilizer}{two types of fertilizers (‘Fert1’ and ‘Fert2’)}
  \item{Site}{ the experiment was done in two different geographical areas (‘Site1’ and ‘Site2’)}
}
}
\usage{
Barley
}
\description{
Results from an experiment to see the dependency of the yield of barley
}
\examples{
lm(Yield ~ Variety+Soil, data = Barley)
}
\keyword{datasets}
