% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Beefcarcass_comp}
\alias{Beefcarcass_comp}
\title{Beefcarcass_comp}
\format{
KFactor will be used as response variable. In addition to KFactor, breed, gender and age is registered for all
carcasses. Breed is represented with 3 levels (‘NRF’, i.e. Norwegian Red Cattle, ‘Limousin’ and ‘Holstein’)
in the dataset. ‘Holstein’ is a breed specialized on dairy (milk) production, ‘Limousin’ is specialized for beef
production and ‘NRF’ is a so-called combination breed, i.e. bred for both dairy and beef production. There
are 2 genders, ‘Bull’ (male) and ‘cow’ (female).
}
\usage{
Beefcarcass_comp
}
\description{
The data used in this exercise are real data from cattle carcasses in Norwegian Slaughterhouses. Carcasses
are classified in accordance with the EUROP standard, where confirmation (i.e. the overall form of the
carcass) and fat content are graded at 15 point scale(s). The confirmation class correlates well to the so
called KFactor
}
\examples{
lm(KFactor ~ Breed*Gender, data = Beefcarcasses_comp)
}
\keyword{datasets}
