% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BeerDataNotClever}
\alias{BeerDataNotClever}
\title{BeerDataNotClever}
\format{
They could add 3 different ingredients during the brewing process. The various ingredients are
either added or they are not added. They allied themselves with a third student who was a judge. The judge
quantified the taste of a total of 16 samples from each of the two beer brewers (scale 0-10). The tasting was
done in a totally randomized order.
}
\usage{
BeerDataNotClever
}
\description{
Two students at NMBU were arguing about who is best at brewing beer.
They therefore decided to do an experiment.
This data is from the not-clever student.
}
\examples{
lm(Taste ~ Ingredient1, data = BeerDataNotClever)
}
\keyword{datasets}
