% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EX14_1Montg_data}
\alias{EX14_1Montg_data}
\title{EX14_1Montg_data}
\format{
Data for a total of 3 samples in a nested design.
\describe{
  \item{Purity}{(Real number) Purity of raw material}
  \item{Supplier}{(Text string) Three levels: 'Supplier1', 'Supplier2' and 'Supplier3' }
  \item{Batch}{(Text string) Batch nr. factor with 4 levels nested under supplier. i.e. 12 differnt batches. 'Batch1' 'Batch2''Batch3' and 'Batch4'}
}
}
\usage{
EX14_1Montg_data
}
\description{
Data from Excample 14.1 in "Montgommery, D. Dessign and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
lm(Purity ~ Supplier, data = EX14_1Montg_data)
}
\keyword{datasets}
