% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example14.16}
\alias{example14.16}
\title{example14.16}
\format{
A paper manufacturer interested in the effect of three different pulp preparation methods and four
different cooking temperatures on the tensile strength of the paper. Data showed the results of 36 runs in a
factorial experiment to test the tensile strength of the paper.
\describe{
  \item{Temperature}{4 different cooking temperatures for the pulp (200, 225, 250 and 275 F degree)}
  \item{pulp_prep_method}{3 different pulp preparation methods (1, 2 and 3)}
  \item{replicate}{3 replicates to obtain a total of 36 observations}
}
}
\usage{
example14.16
}
\description{
Data from the split-plot design example 14.16 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example14.16)
}
\keyword{datasets}
