% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3.1}
\alias{example3.1}
\title{example3.1}
\format{
The engineer is interested in determining if the RF power (in W) setting
affects the etch rate (in Å/min) by running an experiment of four levels of RF power
with five replicates.
}
\usage{
example3.1
}
\description{
Data from Example 3.1 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example3.1)
}
\keyword{datasets}
