% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example3.10}
\alias{example3.10}
\title{example3.10}
\format{
A textile company weaves a fabric on a large number of looms. It would like
the looms to be homogeneous so that it obtains a fabric of uniform strength. The
engineer suspects that there may be significant variations in strength between looms.
 The engineer selects four looms at random with 4 replicates in each loom.
}
\usage{
example3.10
}
\description{
Data from Example 3.10 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example3.10)
}
\keyword{datasets}
