% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example4.1}
\alias{example4.1}
\title{example4.1}
\format{
Data of 24 samples of the percentage of tubes in the production run without any flicks.
\describe{
  \item{Yield}{(real number) percentage of tubes without any flicks}
  \item{Extrusion_pressure}{(real number) the extrusion pressure
   to extrude billets of PTFE resin combined with a lubricant into tubes
   (in PSI with 4 levels: 8500, 8700, 8900, 9100)}
  \item{Batch}{(real number) the resin is manufactured by an external
  supplier and is delivered to the medical device manufacturer in batches (1,2,3,4,5,6)}
}
}
\usage{
example4.1
}
\description{
Data from Example 4.1 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example4.1)
}
\keyword{datasets}
