% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example5.1}
\alias{example5.1}
\title{example5.1}
\format{
Data of 36 samples of effective life (in hours) observed in the battery
\describe{
  \item{Effective_life}{(real numbers): life of the battery in hours}
  \item{Temperature}{(factor): three levels of temperature in F degree: 15, 75 and 125}
  \item{Material_type}{(factor): three sets of material plate (1, 2 and 3)}
}
}
\usage{
example5.1
}
\description{
Data from Example 5.1 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example5.1)
}
\keyword{datasets}
