% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example5.2}
\alias{example5.2}
\title{example5.2}
\format{
The impurity present in a chemical product is affected by two factors:
pressure (25, 30, 35, 40, 45) and temperature (100, 125, 150). Data showed a
factorial experiment with a single replicate.
}
\usage{
example5.2
}
\description{
Data from Example 5.2 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example5.2)
}
\keyword{datasets}
