% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example5.3}
\alias{example5.3}
\title{example5.3}
\format{
A soft drink bottler is interested in obtaining more uniform fill heights
in the bottles. The data showed 24 samples of the average deviation from the target
fill height. The process engineer can control three variables during the filling process.
\describe{
  \item{percent_carbonation}{three levels of 10, 12 and 14}
  \item{operating_pressure}{two levels of 25 and 30 psi}
  \item{line_speed}{two levels of 200 and 250 bpm}
}
}
\usage{
example5.3
}
\description{
Data from Example 5.3 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example5.3)
}
\keyword{datasets}
