% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example5.6}
\alias{example5.6}
\title{example5.6}
\format{
An engineer is studying methods for improving the ability to detect targets
on a radar scope. Data showed 24 samples of the intensity level at detection.
\describe{
  \item{ground_clutter}{the amount of background noice in 3 levels of low, medium and high}
  \item{filter_type}{type of filter placed over the screen (2 levels: 1 and 2)}
  \item{operators}{4 different operators served as blocks (1, 2, 3 and 4)}
}
}
\usage{
example5.6
}
\description{
Data from Example 5.6 in Montgomery D. 2019. Design and Analysis of Experiments.
9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
summary(example5.6)
}
\keyword{datasets}
