% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{FishingExperimentwPersons}
\alias{FishingExperimentwPersons}
\title{FishingExperimentwPersons}
\format{
Data for a total of 48 samples in a three factor fully crossed design.
\describe{
  \item{Yield}{the yield (in hg) of trout after 4 hours fishing}
  \item{Lake}{4 lakes (‘Lake1’, ‘Lake2’, ‘Lake3’ and ‘Lake4’)}
  \item{Hook}{two hooks (a ‘lure’ and a ‘wobler’)}
  \item{Time}{time of the day ('Morning' and 'Evening')}
  \item{Person}{(Text string) Three levels: Person1, Person2, Person3}
}
}
\usage{
FishingExperimentwPersons
}
\description{
Simulated data used for compulsory assignment in STAT210 H20. Trout fishing yield in four different lakes
}
\examples{
summary(FishingExperimentwPersons)
}
\keyword{datasets}
