% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{LambData}
\alias{LambData}
\title{LambData}
\format{
The data used in this exercise is real data from "Sauekontrollen", a database administered by Animalia AS, more info might be found at https://www.animalia.no/no/Dyr/husdyrkontrollene/sauekontrollen/ (in Norwegian).
Data contains three variables "Farm" (Factor with 6 levels, but they represent 30 different farms), factor "Breed" with 5 levels, and "Carcase_weight" (real numbers in kg). The carcase weight is the carcase weight in kg as registered in the slaughterhouse.
}
\usage{
LambData
}
\description{
The data used in this exercise are real data from sheep carcasses in Norwegian Slaughterhouses.
}
\examples{
lm(Carcase_weight ~ Breed, data = LambData)
}
\keyword{datasets}
