% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ListeriaStore}
\alias{ListeriaStore}
\title{ListeriaStore}
\format{
Data has a total of 48 samples
\describe{
  \item{ListeriaNumber}{(Real number) Listeria density (cfu/g)}
  \item{HamTopping}{(Text string) 6 levels (brands): 'HamT1', 'HamT2', 'HamT3', 'HamT4', 'HamT5', 'HamT6'}
  \item{GroceryStore}{(Text string) 4 levels (Grocery stores): 'Store1', 'Store2', 'Store3', 'Store4'}
}
}
\usage{
ListeriaStore
}
\description{
Simulated data used for compulsory assignment in STAT210 H20. Measurements of Listeria on cold cuts of ham.
}
\examples{
lm(ListeriaNumber ~ HamTopping, data = ListeriaStore)
}
\keyword{datasets}
