% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PortlandCement}
\alias{PortlandCement}
\title{PortlandCement}
\format{
Data for a total of 20 samples of mortar of two different types 'Modified' and 'Unmodified'.
\describe{
  \item{Modified}{(Real number) Bondstrength for modified mortar (in kgf/cm2)}
  \item{Unmodified}{(Real number) Bondstrength for unmodified mortar (in kgf/cm2)}
}
}
\usage{
PortlandCement
}
\description{
Data for the Portland Cement example in Montgommery D. 2019. Design and Analysis of Experiments. 9th Edition. EMEA Edition. John Wiley & Sons UK
}
\examples{
PortlandCement <- stack(PortlandCement)
names(PortlandCement) <- c('Strength','Modification')
lm(Strength ~ Modification, data = PortlandCement)
}
\keyword{datasets}
