% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.22}
\alias{problem2.22}
\title{problem2.22}
\format{
The shelf life of a carbonated beverage is of interest.
Ten bottles are randomly selected and tested and the data are the shelf life (in days)
of each bottle
}
\usage{
problem2.22
}
\description{
Data from Exercise 2.22 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.22)
}
\keyword{datasets}
