% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.24}
\alias{problem2.24}
\title{problem2.24}
\format{
The time to repair an electronic instrument is a normally distributed random variable measured in hours.
The repair times for 16 such instruments were chosen at random.
}
\usage{
problem2.24
}
\description{
Data from Exercise 2.24 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.24)
}
\keyword{datasets}
