% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.26}
\alias{problem2.26}
\title{problem2.26}
\format{
Two machines are used for filling plastic bottles with a net volume of 16 ounces.
The filling processes can be assumed to be normal, with standard deviation of 0.015 and 0.018.
The quality engineering department suspects that both machines fill to the same net volume, whether or not this volume is 16 ounces.
An experiment is performed by taking a random sample from the output of each machine.
}
\usage{
problem2.26
}
\description{
Data from Exercise 2.26 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.26)
}
\keyword{datasets}
