% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.28}
\alias{problem2.28}
\title{problem2.28}
\format{
The data are the burning times (in minutes) of chemical flares of two different formulations.
The design engineers are interested in both the mean and variances of the burning times.
}
\usage{
problem2.28
}
\description{
Data from Exercise 2.28 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.28)
}
\keyword{datasets}
