% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.29}
\alias{problem2.29}
\title{problem2.29}
\format{
An article in Solid State Technology, "Orthogonal Design for Process Optimization
and Its Application to Plasma Etching" by G.Z. Yin and D.W. Jillie (May 1987)
describes an experiment to determine the effect of the C2F6 flow rate on the
uniformity of the etch on a silicon wafer used in integrated circuit manufacturing.
All of the runs were made in random order. Data are the uniformity observation of the two flow rates.
}
\usage{
problem2.29
}
\description{
Data from Exercise 2.29 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.29)
}
\keyword{datasets}
