% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.31}
\alias{problem2.31}
\title{problem2.31}
\format{
Photoresist is a light0sensitive material applied to semi-conductor
wafers so that the circuit pattern can be imaged on to the wafer. After application,
the coated wafers are baked to remove the solvent in the photoresist mixture and to harden the resist.
Data are the measuremnets of photoresist thickness (in kA) for eight wafers baked at two different temperatures.
Assume that all of the runs were made in random order.
}
\usage{
problem2.31
}
\description{
Data from Exercise 2.31 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.31)
}
\keyword{datasets}
