% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.32}
\alias{problem2.32}
\title{problem2.32}
\format{
Front housings for cellphones are manufactured in an injection molding process.
The time the part is allowed to cool in the mold before removal is thought to influence
the occurrence of a particularly troublesome cosmetic defect, flow lines, in the finished
housing. After manufacturing, the housings are inspected visually and assigned a score between
1 and 10 based on their appearance, with 10 corresponding to a perfect part and 1 corresponding to a completely defective part.
An experiment was conducted using two cool-down times, 10 and 20 seconds, and 20 housings were evaluated at each level of cool-down time.
All 40 observations in this experiment were run in random order.
}
\usage{
problem2.32
}
\description{
Data from Exercise 2.32 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.32)
}
\keyword{datasets}
