% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.33}
\alias{problem2.33}
\title{problem2.33}
\format{
Twenty observaions on etch uniformity on cilicon wafers are taken during a qualification experiment for a plasma etcher.
}
\usage{
problem2.33
}
\description{
Data from Exercise 2.33 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.33)
}
\keyword{datasets}
