% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.34}
\alias{problem2.34}
\title{problem2.34}
\format{
The diameter of ball bearing was measured by 12 inspectors, each using two different kinds of calipers.
}
\usage{
problem2.34
}
\description{
Data from Exercise 2.34 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.34)
}
\keyword{datasets}
