% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.35}
\alias{problem2.35}
\title{problem2.35}
\format{
An article observed that monozygotic twins share numerous physical, psychological and pathological traits.
The investigators measured an intelligence score of 10 pairs of twins.
}
\usage{
problem2.35
}
\description{
Data from Exercise 2.35 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.35)
}
\keyword{datasets}
