% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.36}
\alias{problem2.36}
\title{problem2.36}
\format{
An article compares several procedures for predicting the shear strength for steel plate girders.
The data show nine girders in the form of the ratio of predicted to observed load for two of these procedures, the Karlsruhe and Lehigh methods.
}
\usage{
problem2.36
}
\description{
Data from Exercise 2.36 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.36)
}
\keyword{datasets}
