% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.37}
\alias{problem2.37}
\title{problem2.37}
\format{
The deflection temperature under load for two different formulations of ABS
plastic pipe is being studied. Two samples of 12 observations each are prepared using
each formulation and the deflection temperatures (in F degree) are reported.
}
\usage{
problem2.37
}
\description{
Data from Exercise 2.37 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.37)
}
\keyword{datasets}
