% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.39}
\alias{problem2.39}
\title{problem2.39}
\format{
In semiconductor manufacturing, wet chemical etching is often used to
remove silicon from the backs of wafers prior to metalization, The etch rate is an important
characteristic of this process. two different etching solutions are being evaluated. Eight
randomly selected wafers have been etched in each solution, and the observed etch rates (in mils/min)
are reported.
}
\usage{
problem2.39
}
\description{
Data from Exercise 2.39 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.39)
}
\keyword{datasets}
