% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.45}
\alias{problem2.45}
\title{problem2.45}
\format{
Data for developing equation 2.50 from textbook
}
\usage{
problem2.45
}
\description{
Data from Exercise 2.45 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.45)
}
\keyword{datasets}
