% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem2.46}
\alias{problem2.46}
\title{problem2.46}
\format{
Data for developing equation for finding confidence interval on the difference in the
means of two normal distributions where variances are different.
}
\usage{
problem2.46
}
\description{
Data from Exercise 2.46 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem2.46)
}
\keyword{datasets}
