% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.10}
\alias{problem3.10}
\title{problem3.10}
\format{
An article described a randomized, double-blind, sham-controlled, feasibility
and dosing study to determine if a common pulsing electromagnetic field (PEMF) treatment
could moderate the substantial osteopenia that occurs after forearm disuse. Subjects were
randomized into 4 groups after a distal radius fracture, or carpal surgery requiring immobilization
in a cast. Active or identical sham PEMF transducers were worn on the distal forearm for 1, 2, or 4 h/day
for 8 weeks starting after cast removal ("baseline") when bon density continues to decline. Bone mineral
density (BMD) and bone geometry were measured in the distal forearm by dual energy X-ray absortiometry (DXA)
and peripheral quantitative computed tomography (pQCT). The data are the percent losses in BMD measurements on
the radius after 16 weeks for patients wearing the active or sham PEMF transducers for 1, 2, or 4 h/day (data
were constructed to match the means and standard deviations read from a graph in the paper)
}
\usage{
problem3.10
}
\description{
Data from Exercise 3.10 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.10)
}
\keyword{datasets}
