% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.11}
\alias{problem3.11}
\title{problem3.11}
\format{
The tensile strength of Portland cement is being studied. 4 different mixing techniques
can be used economically. A completely randomized experiment was conducted.
}
\usage{
problem3.11
}
\description{
Data from Exercise 3.11 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.11)
}
\keyword{datasets}
