% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.14}
\alias{problem3.14}
\title{problem3.14}
\format{
A product developer is investigating the tensile strength of a new synthetic
fiber that will be used to make cloth for men's shirts. Strength is usually affected
by the percentage of cotton used in the blend of materials for the fiber. The engineer
conducts a completely randomized experiment with five levels of cotton content and replicates
the experiment 5 times.
}
\usage{
problem3.14
}
\description{
Data from Exercise 3.14 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.14)
}
\keyword{datasets}
