% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.16}
\alias{problem3.16}
\title{problem3.16}
\format{
A pharmaceutical manufacturer wants to investigate the bioactivity of a new drug. A completely
randomized single-factor experiment was conducted with three dosage levels, and the following results were obtained.
}
\usage{
problem3.16
}
\description{
Data from Exercise 3.16 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.16)
}
\keyword{datasets}
