% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.17}
\alias{problem3.17}
\title{problem3.17}
\format{
A rental care company wants to investigate whether the type of car rented affects the
length of the rental period. An experiment is run for one week at a particular location, and 10
rental contracts are selecyed at random for each car type.
}
\usage{
problem3.17
}
\description{
Data from Exercise 3.17 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.17)
}
\keyword{datasets}
