% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.18}
\alias{problem3.18}
\title{problem3.18}
\format{
Data are the scores of golf play divided by the golf season in the year:
summer, winter and shoulder. The player wanted to test his hypothesis on which
season he achieved the best golf play.
}
\usage{
problem3.18
}
\description{
Data from Exercise 3.18 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.18)
}
\keyword{datasets}
