% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.19}
\alias{problem3.19}
\title{problem3.19}
\format{
A regional opera company has tried 3 approaches to solicit donations from 24 potential
sponsors. The sponsors were randomly divided into three groups of eight and one approach was used for each group.
The data showed the dollar amounts of the resulting contributions.
}
\usage{
problem3.19
}
\description{
Data from Exercise 3.19 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.19)
}
\keyword{datasets}
