% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.20}
\alias{problem3.20}
\title{problem3.20}
\format{
An experiment was run to determine whether 4 specific firing temperatures affect the
density of a certain type of brick. The temperatures are: 100, 125, 150 and 175.
}
\usage{
problem3.20
}
\description{
Data from Exercise 3.20 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.20)
}
\keyword{datasets}
