% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.22}
\alias{problem3.22}
\title{problem3.22}
\format{
A manufacturer of television sets is interested in the effect on tube conductivity of
four different types of coating for color picture tubes. A complete randomized experiment was
conducted.
}
\usage{
problem3.22
}
\description{
Data from Exercise 3.22 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.22)
}
\keyword{datasets}
