% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{problem3.26}
\alias{problem3.26}
\title{problem3.26}
\format{
The response time in milliseconds was determined for 3 different types of circuits that
could be used in an automatic valve shutoff mechanism.
}
\usage{
problem3.26
}
\description{
Data from Exercise 3.26 in "Montgomery, D. Design and Analysis of Experiments. 9th Ed. Wiley."
}
\examples{
summary(problem3.26)
}
\keyword{datasets}
