% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{RifleTokyoOL}
\alias{RifleTokyoOL}
\title{RifleTokyoOL}
\format{
#' #' \describe{
  \item{Points}{(real number) The total number of points in a 5 shot series.}
  \item{Style}{(Textstring) Three possible styles}
  \item{Gender}{(Textstring) Two genders: males and female}
  \item{Athlete}{(Textstring) Name of the athlete}
  \item{Simulated}{(Text string) Yes or No.}
}
}
\usage{
RifleTokyoOL
}
\description{
The data to be used are results from Tokyo Olympics, 50-mter rifle in three
positions. The shooters (i.e. athletes) competes in three different shooting
positions (styles), (i) kneeling  ("kne" in Norwegian), (ii) prone ("liggende"
in Norwegian) and (iii) standing position. In the final with 8 athletes, the
athletes shoot three five-shot series in each position, with maximum score
54,5 points in each series.
}
\examples{
lm(Points ~ Style, data = RifleTokyoOL)
}
\keyword{datasets}
