% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SalmonLice}
\alias{SalmonLice}
\title{SalmonLice}
\format{
Data has a total of 48 samples
\describe{
  \item{SalmonLice}{(Real number) Average number of lice that was observed in a cage (merd)}
  \item{Farm}{(Text string) 2 levels: 'Farm1' and 'Farm2'. Note that 'Farm1' under 'Lumpfish' is not the same farm as 'Farm1' under 'Wrappers' and 'Farm1' under control etc. Thus 6 different farms are represented.}
  \item{CleanerFish}{(Text string) Three levels: 'lumpfish' (rognkjeks in Norwegian),  'wrasses' (leppefisk in Norwegian)  and 'control' (i.e. no cleanerfish).}
}
}
\usage{
SalmonLice
}
\description{
Simulated data used for Exercise 3 in Exam in STAT210 2019.
The the average number of salmon lice was observed for 8 fish cages under a total of 6 farms where lumpfish (rognkjeks in Norwegian) was used as cleaner fish in 2 farms
and wrasses (leppefisk in Norwegian)  was used in 2 farms and 2 farms was used as control.
}
\examples{
lm(SalmonLice ~ CleanerFish, data = SalmonLice)
}
\keyword{datasets}
