% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sires}
\alias{sires}
\title{sires}
\format{
Data for a total of 40 from 8 daughters under five sires
\describe{
  \item{sire}{(Integer 1-5) Represent 5 sires; 1- Åge,  2-Hynar,  3-Lars,  4-Buss and 5-Gorm}
  \item{milk}{(Real number) Annual milk yield (l) for daughters of the five sires.}
}
}
\usage{
sires
}
\description{
Data for 5 sires. Used in STAT210 over several years
}
\examples{
lm(milk ~ sire, data = sires)
}
\keyword{datasets}
