% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SoyaData}
\alias{SoyaData}
\title{SoyaData}
\format{
Data for a total of 6 piglets with 3 different diets.
\describe{
  \item{Diet}{(text string) representing 3 different diets: 'NONS' (no soya),  'S1' (a diet with soya) and 'S2' (another diet with soya)}
  \item{Weight}{(real number) weight of piglet in kg)}
}
}
\usage{
SoyaData
}
\description{
Example data created by STAT210 Teacher
}
\examples{
lm(Weight ~ Diet, data = SoyaData)
}
\keyword{datasets}
