% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Trout}
\alias{Trout}
\title{Trout}
\format{
Data for a total of 40 trouts.
\describe{
  \item{Sulfamerazine}{(Text string (2 letters)) Representing 4 different
  levels: S00, S05, S10 and S15 respectively with different levels of
  sulfamerazine i.e. 5, 10, and 15g per 100 pounds of fish.}
  \item{Hemo}{(Real number) Hemoglobin in blood (g/ 100 mlk) after 35 days.}
}
}
\usage{
Trout
}
\description{
Data from excercise 3.15 in "Dean, Voss & Dragulijc. 2017. Design and
Analysis of experiments. 2nd
}
\examples{
summary(Trout)
}
\keyword{datasets}
